Revert "refactor file and class structure in schemas"
This reverts commit 6db89323d39b7a71a039bf7183be5b9b5887ab0d.
