--- athena/athena/schemas/feedback/__init__.py
@@ -1,7 +1,0 @@
-from .graded_feedback import GradedFeedback
-from .graded_feedback import GradedProgrammingFeedback
-from .graded_feedback import GradedModellingFeedback
-from .graded_feedback import GradedTextFeedback
-from .non_graded_feedback import NonGradedFeedback
-from .non_graded_feedback import NonGradedProgrammingFeedback
-from .feedback import Feedback
