--- athena/athena/schemas/feedback/non_graded_feedback/non_graded_feedback.py
@@ -1,5 +1,0 @@
-from athena.schemas.feedback.feedback import Feedback
-
-
-class NonGradedFeedback(Feedback):
-    pass
