+++ athena/athena/schemas/graded_feedback.py
@@ -3,10 +3,10 @@ from typing import Optional
 
 from pydantic import Field
 
-from athena.schemas.schema import Schema
+from .schema import Schema
 
 
-class Feedback(Schema, ABC):
+class GradedFeedback(Schema, ABC):
     id: Optional[int] = Field(None, example=1)
     title: Optional[str] = Field(None,
                                  description="The title of the feedback that is shown to the student.",
@@ -14,6 +14,12 @@ class Feedback(Schema, ABC):
     description: Optional[str] = Field(None,
                                        description="The detailed feedback description that is shown to the student.",
                                        example="Your solution is correct.")
+    credits: float = Field(0.0,
+                           description="The number of points that the student received for this feedback.",
+                           example=1.0)
+    structured_grading_instruction_id: Optional[int] = Field(None,
+                                                             description="The id of the structured grading instruction that this feedback belongs to.",
+                                                             example=1)
 
     meta: dict = Field({}, example={})
